fix: correctly render text from menu items
Before we were creating each menu item's render picture
using the height of the text. This caused the text of
some of the menu items to not be rendered with a
foreground color.

Now we instead create the render pictures with the
height of the entire container for the menu item,
which allows for correct filling of the foreground
when trying to render the text. :)