feat: custom font rendering support (#8)
* chore: added font

* draft: rough implementation of custom font support

We now use JetBrains Mono as the font for rendering
text. This implementation is not yet fully working
because the text is drawn incorrectly. This will be
fixed soon.

* fix: pad text within its box

* fix: init font picture and pixmap once at the start

* fix: fg rendering now works properly :D

* fix: use "dark mode" :)

* fix: clean up mapping to font loading error

* fix: correct foreground filling

* feat: add support for multi line text drawing

* ref: move `RenderString` related items into separate file

* ref: minor cleanup

* fix: make font rendering more robust and flexible

* feat: implement custom font rendering for the menu

* chore: fixing clippy warnings :)

* chore: cleaning up menu related items

* fix: correctly render text from menu items

Before we were creating each menu item's render picture
using the height of the text. This caused the text of
some of the menu items to not be rendered with a
foreground color.

Now we instead create the render pictures with the
height of the entire container for the menu item,
which allows for correct filling of the foreground
when trying to render the text. :)

* chore: fix clippy warnings